+++ onBoardProject/package.json
@@ -26,7 +26,7 @@
     "@react-native-community/eslint-config": "^2.0.0",
     "@types/jest": "^26.0.23",
     "@types/react-native": "^0.64.5",
-    "@types/react-native-vector-icons": "^6.4.8",
+    "@types/react-native-vector-icons": "6.4.8",
     "@types/react-test-renderer": "^16.9.2",
     "babel-jest": "^26.6.3",
     "eslint": "^7.14.0",
